requirements.txt
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
bson==0.5.7
certifi==2018.11.29
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.3
cryptography==2.9.2
DateTime==4.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.8
iniconfig==1.0.1
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
jsonref==0.2
jsonschema==2.6.0
jwt==1.0.0
lazy-object-proxy==1.4.3
logzero==1.5.0
make-response==1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.0.0rc3
mccabe==0.6.1
module==0.0.4
monotonic==1.5
more-itertools==8.5.0
msgpack-python==0.5.6
mysql-connector-python==8.0.21
numpy==1.19.1
packaging==20.4
pandas==1.1.1
passlib==1.7.2
pipenv==2020.8.13
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pycparser==2.20
PyJWT==1.4.2
pylint==2.6.0
pymongo==3.11.0
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
selenium==3.141.0
simplejson==3.17.2
six==1.12.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.25.10
uuid==1.30
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
yum==0.0.1
zope.interface==5.1.0